+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/episode/EpisodeContainerFragment.kt
@@ -108,7 +108,7 @@ class EpisodeContainerFragment :
         get() = args.episodeUuid
 
     private val timestamp: Duration?
-        get() = args.timestamp
+        get() = arguments?.getLong(ARG_TIMESTAMP_IN_SECS, -1)?.takeIf { it >= 0 }?.seconds
 
     private val episodeViewSource: EpisodeViewSource
         get() = args.source